Model Tuning/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
werkzeug==2.3.7
langchain==0.0.267
langchain-openai==0.0.2
langchain-community==0.0.9
openai>=1.6.1
chromadb==0.4.18
pillow==10.0.0
python-dotenv==1.0.0
jinja2==3.1.2
gunicorn==21.2.0
numpy==1.24.4
pandas==2.0.3
matplotlib==3.7.3
scikit-learn==1.3.0
tiktoken>=0.5.2,<0.6.0